requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
pymongo==4.13.2
openai==1.98.0
python-dotenv==1.1.1
bcrypt==4.3.0
requests==2.31.0
langdetect==1.0.9
googletrans==4.0.0rc1
Werkzeug==3.1.3
dnspython==2.7.0